+++ src/PhpPresentation/Writer/ODPresentation/ObjectsChart.php
@@ -183,9 +183,11 @@ class ObjectsChart extends AbstractDecoratorWriter
         $this->xmlContent->writeAttribute('chart:style-name', 'styleChart');
         $this->xmlContent->writeAttributeIf($chartType instanceof Area, 'chart:class', 'chart:area');
         $this->xmlContent->writeAttributeIf($chartType instanceof AbstractTypeBar, 'chart:class', 'chart:bar');
+        if (!($chartType instanceof Doughnut)) {
+            $this->xmlContent->writeAttributeIf($chartType instanceof AbstractTypePie, 'chart:class', 'chart:circle');
+        }
         $this->xmlContent->writeAttributeIf($chartType instanceof Doughnut, 'chart:class', 'chart:ring');
         $this->xmlContent->writeAttributeIf($chartType instanceof Line, 'chart:class', 'chart:line');
-        $this->xmlContent->writeAttributeIf($chartType instanceof AbstractTypePie, 'chart:class', 'chart:circle');
         $this->xmlContent->writeAttributeIf($chartType instanceof Scatter, 'chart:class', 'chart:scatter');
 
         //**** Title ****
@@ -439,13 +441,33 @@ class ObjectsChart extends AbstractDecoratorWriter
     {
         // chart:legend
         $this->xmlContent->startElement('chart:legend');
-        $this->xmlContent->writeAttribute('chart:legend-position', 'end');
+        switch ($chart->getLegend()->getPosition()) {
+            case Chart\Legend::POSITION_BOTTOM:
+                $position = 'bottom';
+                break;
+            case Chart\Legend::POSITION_LEFT:
+                $position = 'start';
+                break;
+            case Chart\Legend::POSITION_TOP:
+                $position = 'top';
+                break;
+            case Chart\Legend::POSITION_TOPRIGHT:
+                $position = 'top-end';
+                break;
+            case Chart\Legend::POSITION_RIGHT:
+            default:
+                $position = 'end';
+                break;
+        }
+        $this->xmlContent->writeAttribute('chart:legend-position', $position);
         $this->xmlContent->writeAttribute('svg:x', Text::numberFormat(CommonDrawing::pixelsToCentimeters($chart->getLegend()->getOffsetX()), 3) . 'cm');
         $this->xmlContent->writeAttribute('svg:y', Text::numberFormat(CommonDrawing::pixelsToCentimeters($chart->getLegend()->getOffsetY()), 3) . 'cm');
         $this->xmlContent->writeAttribute('style:legend-expansion', 'high');
         $this->xmlContent->writeAttribute('chart:style-name', 'styleLegend');
         // > chart:legend
         $this->xmlContent->endElement();
+
+
     }
 
     /**
@@ -457,6 +479,11 @@ class ObjectsChart extends AbstractDecoratorWriter
         $this->xmlContent->startElement('style:style');
         $this->xmlContent->writeAttribute('style:name', 'styleLegend');
         $this->xmlContent->writeAttribute('style:family', 'chart');
+        // style:chart-properties
+        $this->xmlContent->startElement('style:chart-properties');
+        $this->xmlContent->writeAttribute('chart:auto-position', 'true');
+        // > style:chart-properties
+        $this->xmlContent->endElement();
         // style:text-properties
         $this->xmlContent->startElement('style:text-properties');
         $this->xmlContent->writeAttribute('fo:color', '#'.$chart->getLegend()->getFont()->getColor()->getRGB());
@@ -714,6 +741,19 @@ class ObjectsChart extends AbstractDecoratorWriter
                 $this->xmlContent->writeAttribute('chart:symbol-height', $symbolSize.'cm');
             }
         }
+
+        $separator = $series->getSeparator();
+        if (!empty($separator)) {
+            // style:chart-properties/chart:label-separator
+            $this->xmlContent->startElement('chart:label-separator');
+            if ($separator == PHP_EOL) {
+                $this->xmlContent->writeRaw('<text:p><text:line-break /></text:p>');
+            } else {
+                $this->xmlContent->writeElement('text:p', $separator);
+            }
+            $this->xmlContent->endElement();
+        }
+
         // > style:chart-properties
         $this->xmlContent->endElement();
         // style:graphic-properties