+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Changes
 
 ### Features
+- ODPresentation Writer : Support for the position of Legend - @Progi1984 GH-355
+- ODPresentation Writer : Support for DoughnutChart - @Progi1984 GH-355
 - PowerPoint2007 Writer : Support for DoughnutChart - @Progi1984 GH-355
 
 ## 0.9.0 - 2017-07-05